+++ Editor/Weaver/WeaverAttributes.cs
@@ -1,5 +1,4 @@
 ﻿using Mono.Cecil;
-using System;
 using UnityEngine;
 
 namespace Hertzole.CecilAttributes.Editor
@@ -8,6 +7,7 @@ namespace Hertzole.CecilAttributes.Editor
     {
         private static readonly BaseProcessor[] processors = new BaseProcessor[]
         {
+            new LogCalledProcessor(),
             new ResetStaticProcessor()
         };
 
@@ -21,31 +21,27 @@ namespace Hertzole.CecilAttributes.Editor
                 {
                     if (type.Name != "<Module>")
                     {
-                        Type realType = Type.GetType(type.FullName + ", " + type.Module.Assembly.FullName);
-                        if (realType != null)
+                        for (int i = 0; i < processors.Length; i++)
                         {
-                            for (int i = 0; i < processors.Length; i++)
+                            if (processors[i].NeedsMonoBehaviour && !type.IsSubclassOf<MonoBehaviour>())
                             {
-                                if (processors[i].NeedsMonoBehaviour && !realType.IsSubclassOf(typeof(MonoBehaviour)))
-                                {
-                                    continue;
-                                }
+                                continue;
+                            }
 
-                                if (!processors[i].IsValidClass(type))
-                                {
-                                    continue;
-                                }
+                            if (!processors[i].IsValidClass(type))
+                            {
+                                continue;
+                            }
 
-                                (bool success, bool dirtyClass) = processors[i].ProcessClass(module, type, realType);
-                                if (dirtyClass)
-                                {
-                                    dirty = true;
-                                }
+                            (bool success, bool dirtyClass) = processors[i].ProcessClass(module, type);
+                            if (dirtyClass)
+                            {
+                                dirty = true;
+                            }
 
-                                if (!success)
-                                {
-                                    return (false, false);
-                                }
+                            if (!success)
+                            {
+                                return (false, false);
                             }
                         }
                     }